backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
python-multipart==0.0.6
twilio==8.10.0
openai==1.3.7
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
python-dotenv==1.0.0
aiofiles==23.2.1
httpx==0.25.2
instaloader==4.10.3
